src/main/java/com/halfgallon/withcon/domain/member/controller/MemberController.java
package com.halfgallon.withcon.domain.member.controller;

import com.halfgallon.withcon.domain.auth.security.service.CustomUserDetails;
import com.halfgallon.withcon.domain.member.dto.request.UpdateMemberRequest;
import com.halfgallon.withcon.domain.member.service.MemberService;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.annotation.AuthenticationPrincipal;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PatchMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequiredArgsConstructor
@RequestMapping("/member")
public class MemberController {

  private final MemberService memberService;

  @GetMapping("/me")
  public ResponseEntity<?> findMyInformation(
      @AuthenticationPrincipal CustomUserDetails userDetails) {
    return ResponseEntity.ok(memberService.getMyInformation(userDetails.getId()));
  }

  @PatchMapping
  public ResponseEntity<Void> updateMember(
      @AuthenticationPrincipal CustomUserDetails userDetails,
      @RequestBody UpdateMemberRequest updateMemberRequest) {
    memberService.updateMember(userDetails.getId(), updateMemberRequest);
    return ResponseEntity.ok().build();
  }

  @DeleteMapping
  public ResponseEntity<Void> deleteMember(@AuthenticationPrincipal CustomUserDetails userDetails) {
    memberService.deleteMember(userDetails.getId());
    return ResponseEntity.ok().build();
  }
}
